+++ src/main/java/org/takesome/kaylasEngine/gui/adapters/FrameAttributesLoader.java
@@ -1,7 +1,0 @@
-package org.takesome.kaylasEngine.gui.adapters;
-
-import org.takesome.kaylasEngine.gui.components.Attributes;
-
-public interface FrameAttributesLoader {
-    Attributes getAttributes(String framePath);
-}